+++ arcade_machine/main.py
@@ -1,4 +1,4 @@
-from pygame import QUIT, K_ESCAPE
+from pygame import QUIT, K_ESCAPE, KEYDOWN, mouse
 from pygame import init as pygame_init
 from pygame import quit as pygame_quit
 from pygame import FULLSCREEN
@@ -10,6 +10,9 @@ from arcade_machine.events import CHANGE_GAME
 
 from games.main_menu import MainMenu
 from games.pong import Pong
+from games.arcade_settings import ArcadeSettings
+
+from arcade_machine.components.music_player import stop_music
 
 SCREEN_WIDTH = 1024
 SCREEN_HEIGHT = 768
@@ -17,7 +20,8 @@ SCREEN_HEIGHT = 768
 def main():
     pygame_init()
     clock = Clock()
-    screen = pygame_set_display_mode((SCREEN_WIDTH, SCREEN_HEIGHT), FULLSCREEN)
+    screen = pygame_set_display_mode((SCREEN_WIDTH, SCREEN_HEIGHT))#, FULLSCREEN) TODO: Implement FULLSCREEN
+    mouse.set_visible(False)
 
     current_game = None
     current_game_title = "MainMenu"
@@ -25,6 +29,7 @@ def main():
 
     while True:
         if game_changed:
+            stop_music() # Stop music to prevent sound mix issues when initializing the next game
             current_game = get_game(current_game_title)
             current_game.initialize()
             game_changed = False
@@ -36,9 +41,6 @@ def main():
             elif event.type == CHANGE_GAME:
                 game_changed = True
                 current_game_title = event.game
-            elif event.key == K_ESCAPE:
-                pygame_quit()
-                exit()
             else:
                 current_game.handle_event(event)
         
@@ -52,7 +54,8 @@ def main():
 def get_game(game_title):
     games = {
         "MainMenu": MainMenu(),
-        "Pong": Pong()
+        "Pong": Pong(),
+        "Settings": ArcadeSettings()
     }
     return games[game_title]