+++ CHANGELOG.md
@@ -5,3 +5,11 @@
  - created main event loop
  - created main menu
  - created dummy pong game
+
+## [0.0.1]
+ - improved the appearance of the Main Menu and setup a 'Title' class for game title appearance attributes
+ - created Arcade Settings game for users to change the system volume and quit the console
+ - created a color blender for animated backgrounds that change color
+ - created a music player component to manage system Music
+ - created a sound mixer component to manage system Sounds
+ - added fonts and images to resources for demonstrations